#124: Tidy up
